+++ src/options/list.rs
@@ -4,13 +4,13 @@
 use colored::*;
 use std::collections::HashSet;
 // Local imports
-use crate::utilities::{find_version, get_current_changesets, update_version};
+use crate::utilities::{find_largest_version, get_current_changesets};
 
 pub fn list_changesets() {
     // Get the changesets and list them
     let changesets = get_current_changesets();
     // Find the current project version
-    let new_version = update_version(&changesets, find_version());
+    let new_version = find_largest_version(&changesets).unwrap();
     // Print the new version to set with these changesets
     println!("# New version to be bumped: v{}.\n", new_version.blue());
     // Add a vec of tags that were visited
@@ -20,10 +20,7 @@ pub fn list_changesets() {
         // Print the first layer
 
         // Filter the changesets for all those that match the change type
-        for changeset in changesets
-            .iter()
-            .filter(|c| c.change == format!("\t{}", change_type))
-        {
+        for changeset in changesets.iter().filter(|c| c.change.trim() == change_type) {
             // If this tag has been printed already, then continue
             if printed_tags.contains(&changeset.tag) {
                 continue;
@@ -33,15 +30,15 @@ pub fn list_changesets() {
             // And now, filter all the changesets for the tag and the change type
             for nested_changeset in changesets
                 .iter()
-                .filter(|c| c.change == format!("\t{}", change_type) && c.tag == changeset.tag)
+                .filter(|c| c.change.trim() == change_type && c.tag == changeset.tag)
             {
                 // If this changeset has a module, include it. If not, then just don't
-                if nested_changeset.module == "" {
+                if nested_changeset.modules == "" {
                     println!("    - {}", nested_changeset.message);
                 } else {
                     println!(
                         "    - {}: {}",
-                        nested_changeset.module.blue(),
+                        nested_changeset.modules.blue(),
                         nested_changeset.message
                     );
                 }