+++ src/utilities/changelog_utils.rs
@@ -24,7 +24,7 @@ pub fn open_changelog() -> Vec<String> {
 }
 
 pub fn create_changelog(content: Vec<String>, version: &String) {
-    // Crea un nuevo archivo CHANGELOG.md
+    // Create a new CHANGELOG.md file
     let mut file = fs::File::create("CHANGELOG.md").expect("Error creating the CHANGELOG.md");
 
     // Write the entire CHANGELOG content
@@ -54,12 +54,12 @@ pub fn new_changelog_entry(changesets: &Vec<Changeset>, version: &String) -> Vec
         // Filter for all the same tags
         for nested_changeset in changesets.iter().filter(|c| c.tag == changeset.tag) {
             // Then, write all the changes
-            if nested_changeset.module == "" {
+            if nested_changeset.modules == "" {
                 content.push(format!("- {}.\n", nested_changeset.message));
             } else {
                 content.push(format!(
                     "- {}: {}.\n",
-                    nested_changeset.module, nested_changeset.message
+                    nested_changeset.modules, nested_changeset.message
                 ));
             }
         }