+++ src/utilities/mod.rs
@@ -1,15 +1,23 @@
-mod changelog_utils;
+// Module declarations
+pub mod ai_calls;
+pub mod ai_message_generator;
+pub mod changelog_utils;
+pub mod changeset_structures;
+pub mod version_operations;
+
+// Re-exports
+pub use ai_message_generator::{generate_ai_message, AIConfig};
+pub use changelog_utils::{create_changelog, new_changelog_entry, open_changelog};
+
 /// Make the modules accessible
 mod changesets_utilities;
 mod sets_utils;
 mod subcommands;
-// Import the needed methods
-pub use changelog_utils::{create_changelog, new_changelog_entry, open_changelog};
+// Local imports
+use crate::options::Changeset;
 pub use changesets_utilities::get_current_changesets;
 pub use sets_utils::{create_changeset_folder, write_changeset_file};
 pub use subcommands::create_subcommands;
-// Local imports
-use crate::options::Changeset;
 // Libraries to use
 use regex::Regex;
 use std::fs;
@@ -18,12 +26,12 @@ use toml::Value;
 
 pub fn find_version() -> String {
     // Find the version in the current path
-    let version_path = find_version_in_file();
+    let version_paths = find_version_in_file();
     // Using this, return the version
-    open_path(version_path)
+    open_path(version_paths[0].clone())
 }
 
-pub fn find_version_in_file() -> String {
+pub fn find_version_in_file() -> Vec<String> {
     // Search the `pyproject.toml` in the root folder
     let route = "pyproject.toml";
 
@@ -43,32 +51,34 @@ pub fn find_version_in_file() -> String {
         }
     };
 
-    // Search the [tool.sempyver] version path
-    let version_path: String;
+    // Search the [tool.semverai] version path
+    let mut version_paths: Vec<String> = Vec::new();
     if let Some(tool) = toml_config.get("tool") {
-        if let Some(sempyver) = tool.get("sempyver") {
-            if let Some(possible_path) = sempyver.get("version_path") {
-                if let Some(path) = possible_path.get(0) {
-                    version_path = path.to_string().replace("\"", "");
+        if let Some(semverai) = tool.get("semverai") {
+            if let Some(possible_paths) = semverai.get("version_path") {
+                if let Some(paths) = possible_paths.as_array() {
+                    for path in paths {
+                        version_paths.push(path.to_string().replace("\"", ""));
+                    }
                 } else {
                     panic!("The version path doesn't include a path");
                 }
             } else {
-                panic!("The sempyver utility doesn't include a `version_path` field")
+                panic!("The semverai utility doesn't include a `version_path` field")
             }
         } else {
             panic!(
-                "The pyproject doesn't have a sempyver as tool. You should have [tool.sempyver]."
+                "The pyproject doesn't have a semverai as tool. You should have [tool.semverai]."
             )
         }
     } else {
-        panic!("The pyproject doesn't have tools associated. Please add the `sempyver` tool as [tool.sempyver].")
+        panic!("The pyproject doesn't have tools associated. Please add the `semverai` tool as [tool.semverai].")
     }
-    if version_path == "" {
-        panic!("Couldn't find the version in the provided path.")
+    if version_paths.is_empty() {
+        panic!("Couldn't find any version paths in the configuration.")
     }
-    // Return the version path
-    version_path
+    // Return the version paths
+    version_paths
 }
 
 pub fn open_path(path: String) -> String {
@@ -109,72 +119,61 @@ pub fn open_path(path: String) -> String {
     panic!("Couldn't find the version in the path {}. Try with the following version names: [\"version\", \"__version__\"]", path);
 }
 
-pub fn update_version(changes: &Vec<Changeset>, version: String) -> String {
-    // Parse the current version
-    let current_version: Vec<u32> = version
-        .split('.')
-        .map(|s| {
-            s.parse()
-                .unwrap_or_else(|_| panic!("Invalid version: {}", version))
-        })
-        .collect();
+fn update_version_path(new_version: &str) {
+    // Find all version paths
+    let version_paths = find_version_in_file();
+    // Get the current version
+    let current_version = find_version();
 
-    // Find the maximum change type in the list of changes
-    let mut max_change = 'P'; // Default value representing no changes
-    for changeset in changes {
-        if changeset.change.contains("MAJOR") {
-            max_change = 'M';
-            // Since there's no other higher option, we'll choose
-            // this as the maximum change type and break the iteration here
-            break;
-        } else if changeset.change.contains("MINOR") && max_change != 'M' {
-            max_change = 'N'; // Reset to 'N' if no MAJOR change found
-        } else if changeset.change.contains("PATCH") && max_change != 'N' {
-            max_change = 'P';
+    // Update each file
+    for version_path in version_paths {
+        // Open the file
+        let mut file = match fs::File::open(&version_path) {
+            Ok(file) => file,
+            Err(e) => {
+                panic!("Error opening file {}: {}.", version_path, e);
+            }
+        };
+        // Read the content as a String
+        let mut content = String::new();
+        if let Err(e) = file.read_to_string(&mut content) {
+            panic!("Error reading file {}: {}.", version_path, e);
+        }
+        // Substitute the old version for the new version
+        let updated_content = content.replace(&current_version, new_version);
+        // Reopen the file but this time as writing mode
+        file = match fs::File::create(&version_path) {
+            Ok(file) => file,
+            Err(e) => {
+                panic!("Error creating file {}: {}.", version_path, e);
+            }
+        };
+        // Write the new file
+        if let Err(e) = file.write_all(updated_content.as_bytes()) {
+            panic!("Error writing to file {}: {}.", version_path, e);
         }
     }
-    // Update the version based on the maximum change type
-    let mut updated_version = current_version.clone();
-    match max_change {
-        'M' => updated_version[0] += 1,
-        'N' => updated_version[1] += 1,
-        'P' => updated_version[2] += 1,
-        _ => unreachable!(),
-    }
-    // Convert the updated version to a string
-    updated_version
-        .iter()
-        .map(|&v| v.to_string())
-        .collect::<Vec<String>>()
-        .join(".")
 }
 
-fn update_version_path(new_version: &str) {
-    // Find the current version path
-    let version_path = find_version_in_file();
-    // Open the file
-    let mut file = match fs::File::open(&version_path) {
-        Ok(file) => file,
-        Err(e) => {
-            panic!("Error opening file {}: {}.", version_path, e);
-        }
-    };
-    // Read the content as a String
-    let mut content = String::new();
-    if let Err(e) = file.read_to_string(&mut content) {
-        panic!("Error reading file {}: {}.", version_path, e);
-    }
-    // Substitute the old version for the new version
-    let updated_content = content.replace(find_version().as_str(), new_version);
-    // Reopen the file but this time as writing mode
-    file = match fs::File::create(&version_path) {
-        Ok(file) => file,
-        Err(e) => {
-            panic!("Error creating file {}: {}.", version_path, e);
-        }
-    };
-    // Write the new file
-    if let Err(e) = file.write_all(updated_content.as_bytes()) {
-        panic!("Error writing to file {}: {}.", version_path, e);
+/// Find the largest version in a list of changesets
+pub fn find_largest_version(changesets: &[Changeset]) -> Option<String> {
+    changesets
+        .iter()
+        .filter_map(|c| parse_version(&c.version)) // Parse the versions
+        .max() // Obtain the largest version
+        .map(|(major, minor, patch)| format!("{}.{}.{}", major, minor, patch)) // Convert it back to String
+}
+
+/// Parse a version "MAJOR.MINOR.PATCH" into a tuple (u32, u32, u32)
+fn parse_version(version: &str) -> Option<(u32, u32, u32)> {
+    let parts: Vec<u32> = version
+        .split('.') // Divide into parts
+        .filter_map(|p| p.parse().ok()) // Convert to u32
+        .collect();
+
+    if parts.len() == 3 {
+        Some((parts[0], parts[1], parts[2]))
+    } else {
+        Some((0, 0, 0))
     }
 }